coc/requirements.txt
torch>=2.1.0
torchvision>=0.16.0
torchaudio>=2.1.0

numpy>=1.25.0
scipy>=1.11.0
scikit-learn>=1.3.0
tqdm>=4.66.0
Pillow>=10.0.0

timm>=0.9.0
geomloss>=0.4.0
discord.py